refactor: create seperate files for sub components of photo gallery page
